requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
paddleocr==2.9.1
paddlepaddle==3.0.0
pillow==11.1.0
pydantic==2.10.0
aiosqlite==0.20.0
sqlalchemy==2.0.48
alembic==1.18.4
openpyxl==3.1.5
pywebview==6.1
pyinstaller==6.19.0
passlib[argon2]==1.7.4
itsdangerous==2.2.0
pypdfium2==4.30.0
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.28
